requirements.txt
torch>=1.7.0
transformers>=4.0.0
optuna>=2.0.0
scikit-learn>=0.24.0
pandas>=1.2.0
numpy>=1.19.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
tqdm>=4.50.0
wordcloud>=1.8.0

# Testing dependencies
pytest>=6.2.5
coverage>=6.0.0